remove Interfaces project, moved a couple things to different namespace
