--- ServiceFabric.PubSubActors.Interfaces/ServiceFabric.PubSubActors.Interfaces.csproj
@@ -1,40 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <Copyright>2019</Copyright>
-    <AssemblyTitle>ServiceFabric.PubSubActors.Interfaces</AssemblyTitle>
-    <Authors>Loek Duys</Authors>
-    <TargetFrameworks>net452;net46;netstandard2.0</TargetFrameworks>
-    <PlatformTarget>x64</PlatformTarget>
-    <Platform>x64</Platform>
-    <Version>7.6.2</Version>
-    <AssemblyName>ServiceFabric.PubSubActors.Interfaces</AssemblyName>
-    <PackageId>ServiceFabric.PubSubActors.Interfaces</PackageId>
-    <PackageTags>ServiceFabric;Service;Fabric;Actor;Pub;Sub;Reliable</PackageTags>
-    <PackageProjectUrl>https://github.com/loekd/ServiceFabric.PubSubActors</PackageProjectUrl>
-    <PackageLicenseUrl>https://github.com/loekd/ServiceFabric.PubSubActors/blob/master/LICENSE.md</PackageLicenseUrl>
-    <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
-    <RepositoryType>git</RepositoryType>
-    <RepositoryUrl>https://github.com/loekd/ServiceFabric.PubSubActors.git</RepositoryUrl>
-    <GenerateAssemblyTitleAttribute>false</GenerateAssemblyTitleAttribute>
-    <GenerateAssemblyDescriptionAttribute>false</GenerateAssemblyDescriptionAttribute>
-    <GenerateAssemblyConfigurationAttribute>false</GenerateAssemblyConfigurationAttribute>
-    <GenerateAssemblyCompanyAttribute>false</GenerateAssemblyCompanyAttribute>
-    <GenerateAssemblyProductAttribute>false</GenerateAssemblyProductAttribute>
-    <GenerateAssemblyCopyrightAttribute>false</GenerateAssemblyCopyrightAttribute>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <IncludeSymbols>true</IncludeSymbols>
-    <SignAssembly>true</SignAssembly>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)' == 'Release'">
-    <PublicSign Condition=" '$(OS)' != 'Windows_NT' ">true</PublicSign>
-    <DelaySign>False</DelaySign>
-    <SignAssembly>true</SignAssembly>
-    <AssemblyOriginatorKeyFile>..\ServiceFabric.PubSubActors.snk</AssemblyOriginatorKeyFile>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="Microsoft.ServiceFabric.Actors" Version="3.3.638" />
-    <PackageReference Include="Microsoft.ServiceFabric.Services.Remoting" Version="3.3.638" />
-    <PackageReference Include="Microsoft.VisualStudio.Azure.Fabric.MSBuild" Version="1.6.7" />
-    <PackageReference Include="Newtonsoft.Json" Version="12.0.1" />
-  </ItemGroup>
-</Project>
