+++ ServiceFabric.PubSubActors.sln
@@ -12,8 +12,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ServiceFabric.PubSubActors", "ServiceFabric.PubSubActors\ServiceFabric.PubSubActors.csproj", "{266D0C63-57EE-4C0D-873F-B3235B062CF4}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ServiceFabric.PubSubActors.Interfaces", "ServiceFabric.PubSubActors.Interfaces\ServiceFabric.PubSubActors.Interfaces.csproj", "{8E319BE1-AF0A-4683-997F-7577E369850A}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServiceFabric.PubSubActors.Tests", "ServiceFabric.PubSubActors.Tests\ServiceFabric.PubSubActors.Tests.csproj", "{29169DCB-AA96-4B25-8395-32DC00473F34}"
 EndProject
 Global
@@ -32,14 +30,6 @@ Global
 		{266D0C63-57EE-4C0D-873F-B3235B062CF4}.Release|Any CPU.Build.0 = Release|Any CPU
 		{266D0C63-57EE-4C0D-873F-B3235B062CF4}.Release|x64.ActiveCfg = Release|Any CPU
 		{266D0C63-57EE-4C0D-873F-B3235B062CF4}.Release|x64.Build.0 = Release|Any CPU
-		{8E319BE1-AF0A-4683-997F-7577E369850A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8E319BE1-AF0A-4683-997F-7577E369850A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8E319BE1-AF0A-4683-997F-7577E369850A}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{8E319BE1-AF0A-4683-997F-7577E369850A}.Debug|x64.Build.0 = Debug|Any CPU
-		{8E319BE1-AF0A-4683-997F-7577E369850A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8E319BE1-AF0A-4683-997F-7577E369850A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8E319BE1-AF0A-4683-997F-7577E369850A}.Release|x64.ActiveCfg = Release|Any CPU
-		{8E319BE1-AF0A-4683-997F-7577E369850A}.Release|x64.Build.0 = Release|Any CPU
 		{29169DCB-AA96-4B25-8395-32DC00473F34}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{29169DCB-AA96-4B25-8395-32DC00473F34}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{29169DCB-AA96-4B25-8395-32DC00473F34}.Debug|x64.ActiveCfg = Debug|Any CPU
