+++ ServiceFabric.PubSubActors/BrokerService.cs
@@ -1,14 +1,13 @@
 ﻿using Microsoft.ServiceFabric.Data;
 using Microsoft.ServiceFabric.Data.Collections;
 using Microsoft.ServiceFabric.Services.Runtime;
-using ServiceFabric.PubSubActors.Interfaces;
 using ServiceFabric.PubSubActors.State;
 using System;
 using System.Fabric;
 using System.Threading.Tasks;
 using System.Threading;
 using ServiceFabric.PubSubActors.Helpers;
-using ServiceFabric.PubSubActors.SubscriberServices;
+using ServiceFabric.PubSubActors.Subscriber;
 
 namespace ServiceFabric.PubSubActors
 {
