+++ ServiceFabric.PubSubActors/Helpers/BrokerClient.cs
@@ -6,7 +6,6 @@ using System.Threading.Tasks;
 using Microsoft.ServiceFabric.Actors;
 using Microsoft.ServiceFabric.Actors.Runtime;
 using Microsoft.ServiceFabric.Services.Runtime;
-using ServiceFabric.PubSubActors.Interfaces;
 using ServiceFabric.PubSubActors.State;
 
 namespace ServiceFabric.PubSubActors.Helpers
