+++ ServiceFabric.PubSubActors/State/MessageWrapper.cs
@@ -1,11 +1,12 @@
 ﻿using System;
 using System.Reflection;
 using System.Runtime.Serialization;
+using ServiceFabric.PubSubActors.Helpers;
 
-namespace ServiceFabric.PubSubActors.Interfaces
+namespace ServiceFabric.PubSubActors.State
 {
 	/// <summary>
-	/// Generic message format. Contains message CLR type (full name) and serialized payload. If you know the Message Type you can deserialize 
+	/// Generic message format. Contains message CLR type (full name) and serialized payload. If you know the Message Type you can deserialize
 	/// the payload into that object.
 	/// </summary>
 	[DataContract]
@@ -40,7 +41,7 @@ namespace ServiceFabric.PubSubActors.Interfaces
     public static class MessageWrapperExtensions
     {
 		/// <summary>
-		/// Gets or sets the <see cref="IPayloadSerializer"/> to use when setting the <see cref="MessageWrapper.Payload"/>. 
+		/// Gets or sets the <see cref="IPayloadSerializer"/> to use when setting the <see cref="MessageWrapper.Payload"/>.
 		/// Defaults to <see cref="DefaultPayloadSerializer"/> which uses Json.Net.
 		/// </summary>
 		public static IPayloadSerializer PayloadSerializer { get; set; } = new DefaultPayloadSerializer();
@@ -49,7 +50,7 @@ namespace ServiceFabric.PubSubActors.Interfaces
 	    /// Convert the provided <paramref name="message"/> into a <see cref="MessageWrapper"/>
 	    /// </summary>
 	    /// <param name="message"></param>
-	    /// <returns></returns> 
+	    /// <returns></returns>
 	    public static MessageWrapper CreateMessageWrapper(this object message)
         {
             var messageType = message.GetType();
@@ -66,7 +67,7 @@ namespace ServiceFabric.PubSubActors.Interfaces
         /// Convert the provided <paramref name="messageWrapper"/> into an object of type <typeparamref name="TResult"/>
         /// </summary>
         /// <param name="messageWrapper"></param>
-        /// <returns></returns> 
+        /// <returns></returns>
         public static TResult CreateMessage<TResult>(this MessageWrapper messageWrapper)
         {
             var message = (PayloadSerializer ?? new DefaultPayloadSerializer()).Deserialize<TResult>(messageWrapper.Payload);
@@ -77,7 +78,7 @@ namespace ServiceFabric.PubSubActors.Interfaces
         /// Convert the provided <paramref name="messageWrapper"/> into an object of type <see cref="MessageWrapper.MessageType"/>
         /// </summary>
         /// <param name="messageWrapper"></param>
-        /// <returns></returns> 
+        /// <returns></returns>
         public static object CreateMessage(this MessageWrapper messageWrapper)
         {
             var type = Type.GetType(messageWrapper.MessageType, false);
